refactor: move CIVisibilityProtocol to dedicated _protocols.py
Move _CIVisibilityProtocol out of service_registry.py into
ddtrace/internal/ci_visibility/_protocols.py, consistent with the
other Protocol locations. Renamed to CIVisibilityProtocol (public)
since it's now in a dedicated module.

- _protocols.py: new file with CIVisibilityProtocol
- service_registry.py: import Protocol from _protocols.py
- recorder.py: import Protocol from _protocols.py instead of
  service_registry.py
