+++ ddtrace/internal/ci_visibility/recorder.py
@@ -31,6 +31,7 @@ from ddtrace.internal.ci_visibility._api_client import TestManagementSettings
 from ddtrace.internal.ci_visibility._api_client import TestProperties
 from ddtrace.internal.ci_visibility._api_client import TestVisibilityAPISettings
 from ddtrace.internal.ci_visibility._api_client import _TestVisibilityAPIClientBase
+from ddtrace.internal.ci_visibility._protocols import CIVisibilityProtocol
 from ddtrace.internal.ci_visibility.api._module import TestVisibilityModule
 from ddtrace.internal.ci_visibility.api._session import TestVisibilitySession
 from ddtrace.internal.ci_visibility.api._session import TestVisibilitySessionSettings
@@ -52,7 +53,6 @@ from ddtrace.internal.ci_visibility.git_client import METADATA_UPLOAD_STATUS
 from ddtrace.internal.ci_visibility.git_client import CIVisibilityGitClient
 from ddtrace.internal.ci_visibility.git_data import GitData
 from ddtrace.internal.ci_visibility.git_data import get_git_data_from_tags
-from ddtrace.internal.ci_visibility.service_registry import _CIVisibilityProtocol
 from ddtrace.internal.ci_visibility.service_registry import register_ci_visibility_instance
 from ddtrace.internal.ci_visibility.service_registry import unregister_ci_visibility_instance
 from ddtrace.internal.ci_visibility.utils import _get_test_framework_telemetry_name
@@ -152,7 +152,7 @@ class CIVisibilityTracer(Tracer):
         super().__init__(*args, **kwargs)
 
 
-class CIVisibility(Service, _CIVisibilityProtocol):
+class CIVisibility(Service, CIVisibilityProtocol):
     _instance: Optional["CIVisibility"] = None
     enabled = False
